chore: split context building and packages loading
